prevent popover flicker
the mouseenter causes the popover to show, but mouseout causes it to hide. rendering the popover directly under the cursor causes a mouseout event to fire, so the popover constantly flickers between showing and hiding unless you move the cursor away faster than it can render and register the mouseout event. moving the popover a bit solves the issue.